cc: Fix missing std and C++ exception enabling flags
Xcode pre 26's default C++ standard is C++0x, and the Android NDK can be
also set to disable exceptions by default, which is the case of the
GStreamer SDK.

The wrapper uses `auto` and throw, so force enable those. And for
consistency, add MSVC equivalent flags as well.

See for context:

- https://gitlab.freedesktop.org/gstreamer/cerbero/-/commit/973905f6c0fcbb44da00de17524e51388907c14e
- https://gitlab.freedesktop.org/gstreamer/cerbero/-/jobs/84619199#L499
- https://gitlab.freedesktop.org/gstreamer/cerbero/-/jobs/84619203#L477
- https://gitlab.freedesktop.org/gstreamer/cerbero/-/jobs/84619202#L461